+++ src/screens/Home/Bubble.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { View, Text, Image } from "react-native";
+import { View, Text, Image, TouchableOpacity } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { useState, useEffect } from "react";
 
@@ -31,7 +31,8 @@ const Bubble = ({ navigation }) => {
     const userRef = doc(database, "users", auth.currentUser.uid);
     const q = query(
       collection(database, "bubbles"),
-      where("creatorID", "==", userRef)
+      where("creatorID", "==", userRef),
+      where("computerGenerated", "==", false)
     );
 
     const unsubscribe = onSnapshot(q, (querySnapshot) => {
@@ -43,10 +44,10 @@ const Bubble = ({ navigation }) => {
         counter++;
 
         bubblesArray.push({
-          bubbleID: doc.data().bubbleID,
+          bubbleID: doc.id,
           title: doc.data().title,
         });
-
+        console.log(doc.id);
         if (querySnapshot.size == counter) setBubbles(bubblesArray);
       });
     });
@@ -66,10 +67,16 @@ const Bubble = ({ navigation }) => {
             </Text>
           </View>
           <View style={bubbleStyles.headerAddButtonContainer}>
-            <Image
-              source={addBubbleIcon}
-              style={bubbleStyles.headerAddButton}
-            />
+            <TouchableOpacity
+              onPress={() => {
+                navigation.navigate("AddBubble");
+              }}
+            >
+              <Image
+                source={addBubbleIcon}
+                style={bubbleStyles.headerAddButton}
+              />
+            </TouchableOpacity>
           </View>
         </View>
       </View>
@@ -90,10 +97,6 @@ const Bubble = ({ navigation }) => {
               />
             );
           })}
-
-          {/* <MiniBubble bubbleName="DLSU Friends"/>
-          <MiniBubble bubbleName="La Familia"/>
-          <MiniBubble bubbleName="Work"/> */}
         </ScrollView>
       </View>
     </SafeAreaView>