+++ src/screens/Home/Messages.js
@@ -27,8 +27,7 @@ const Messages = ({ navigation }) => {
   const pencilIcon = useIcon("messagePencilIcon");
 
   const [option, setOption] = useState(null);
-  const [customMessageValue, onChangeText] = useState(null);
-  const [name, updateName] = useState("");
+  const [customMessageValue, setCustomMessage] = useState("");
 
   const dictionary = {
     "Open to Chat": "openToChat",
@@ -38,17 +37,22 @@ const Messages = ({ navigation }) => {
   };
 
   useLayoutEffect(() => {
+    const unsubscribeAll = [];
+
     const initialUpdate = async () => {
       const userRef = doc(database, "users", auth.currentUser.uid);
-      const dataSnap = await getDoc(userRef);
-      const dataOSISnap = await getDoc(dataSnap.data().statusID);
+      const user = await getDoc(userRef);
+      const unsubscribe = onSnapshot(user.data().statusID, (indicator) => {
+        setOption(indicator.data().osi);
+        setCustomMessage(indicator.data().message);
+      });
 
-      updateName(dataSnap.data().name);
-      setOption(dataOSISnap.data().osi);
-      onChangeText(dataOSISnap.data().message);
+      unsubscribeAll.push(unsubscribe);
     };
 
     initialUpdate();
+
+    return () => unsubscribeAll.forEach((unsubscribe) => unsubscribe());
   }, []);
 
   useEffect(() => {